Moved IntlContext to context folder
